# modules/samples/swagger-server/src/main/resources/openapi/financial-api-schema.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Financial Data Services API",
        "description": "Apache Axis2 financial data management REST API with comprehensive OpenAPI documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://localhost:8006",
            "description": "Development server"
        }
    ],
    "paths": {
        "/bigdataservice/login": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "User authentication",
                "operationId": "authenticateUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "type": "object",
                "required": [
                    "email",
                    "credentials"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "credentials": {
                        "type": "string",
                        "description": "User password or authentication token"
                    }
                }
            },
            "LoginResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string",
                        "description": "Authentication token"
                    },
                    "expiresIn": {
                        "type": "integer",
                        "description": "Token expiration time in seconds"
                    },
                    "userInfo": {
                        "$ref": "#/components/schemas/UserInfo"
                    },
                    "errorMessage": {
                        "type": "string",
                        "nullable": true
                    }
                }
            },
...
            "UserInfo": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}